Reverse changes from PR 1486 in upstream
We should be allowed to run the bot in "unsupported manners" if we wish, just tell us that it might break things instead of breaking it intentionally. Especially don't exit the entire bot after connecting it to Discord... That causes obvious problems if it's set up as a systemd service. Anyone can see that.